refactor: changed method from GET to POST to send user mail to server
