+++ app.js
@@ -5,8 +5,7 @@ const cookieParser = require("cookie-parser");
 const logger = require("morgan");
 const cors = require("cors");
 
-
-const usersRouter = require("./routes/api/users");
+const usersRouter = require("./routes/users");
 const authRouter = require("./routes/auth");
 
 const app = express();
