+++ routes/auth.js
@@ -8,7 +8,9 @@ const {
 const router = express.Router();
 
 router.post("/login", handleLogin);
-router.get("/logout", handleLogout);
+
+router.post("/logout", handleLogout);
+
 router.get("/refresh", handleRefreshToken);
 
 module.exports = router;
